# docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.StationRequest:
    properties:
      description:
        type: string
      hardware_type:
        type: string
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      name:
        type: string
      station_id:
        type: string
      status:
        type: string
    type: object
  dto.StationResponse:
    properties:
      hardware_type:
        type: string
      id:
        type: integer
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      name:
        type: string
      station_id:
        type: string
      status:
        type: string
    type: object
  dto.UserMeResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  dto.UserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for DEVOPS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BIG-DEVOPS API
  version: "1.0"
paths:
  /v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user via username and password (local) or proxy to
        Keycloak
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user
      tags:
      - Auth
  /v1/me:
    get:
      description: Get details of the currently logged-in user from JWT claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserMeResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /v1/stations:
    get:
      description: Get a list of all EWS Landslide stations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.StationResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all stations
      tags:
      - Stations
    post:
      consumes:
      - application/json
      description: Register a new EWS Landslide station
      parameters:
      - description: Station Data
        in: body
        name: station
        required: true
        schema:
          $ref: '#/definitions/dto.StationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StationResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new station
      tags:
      - Stations
  /v1/stations/{id}:
    delete:
      description: Remove a station from the system (Soft Delete)
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete a station
      tags:
      - Stations
    get:
      description: Get detailed information about a specific station
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StationResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get station by ID
      tags:
      - Stations
    put:
      consumes:
      - application/json
      description: Update station details
      parameters:
      - description: Station ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Station Data
        in: body
        name: station
        required: true
        schema:
          $ref: '#/definitions/dto.StationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StationResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing station
      tags:
      - Stations
  /v1/users:
    get:
      description: Get a list of all users system-wide
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Register a new user with role and hashed password
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /v1/users/{id}:
    delete:
      description: Remove a user from the system (Soft Delete)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      description: Get detailed information about a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user details such as name, email, or role
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing user
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"